feat: change token request mechanism for cheminfo (#27)
Since the `iframe` based solution turned out to be too complicated to maintain (requiring secure cookies and other settings that keep changing as browsers develop) we decided to simply open the token generation page in a new tab. 

This page (implemented and maintained by the cheminfo team), provides besides the token some explanation a "copy to clipboard button". 

Since this token generation page is now always on the same domain as the REST-API of the cheminfo ELN, we no longer need to care about the `cheminfo.org/...` page but only need the domain of the ELN database. 

This commit additionally introduces a small label that links to further documentation of the integration with the cheminfo ELN. 

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>